[step3] 경로 조회 시, 노선에 대한 추가 요금을 확인한다. (long -> int로 변경한다.)

+++ src/main/java/nextstep/subway/applicaion/dto/LineRequest.java
@@ -3,7 +3,7 @@ package nextstep.subway.applicaion.dto;
 public class LineRequest {
     private String name;
     private String color;
-    private Long additionalFare;
+    private int additionalFare;
     private Long upStationId;
     private Long downStationId;
     private int distance;
@@ -33,7 +33,7 @@ public class LineRequest {
         return duration;
     }
 
-    public Long getAdditionalFare() {
+    public int getAdditionalFare() {
         return additionalFare;
     }
 }
